+++ tests/phpunit/MutationTest.php
@@ -37,7 +37,7 @@ namespace Infection\Tests;
 
 use function array_merge;
 use Generator;
-use Infection\EvilNode;
+use Infection\MutatedNode;
 use Infection\Mutation;
 use Infection\Mutator\Arithmetic\Plus;
 use Infection\TestFramework\Coverage\CoverageLineData;
@@ -62,7 +62,7 @@ final class MutationTest extends TestCase
         string $mutatorName,
         array $attributes,
         string $mutatedNodeClass,
-        EvilNode $mutatedNode,
+        MutatedNode $mutatedNode,
         int $mutationByMutatorIndex,
         array $tests,
         array $expectedAttributes,
@@ -109,7 +109,7 @@ final class MutationTest extends TestCase
             Plus::getName(),
             $nominalAttributes,
             Node\Scalar\LNumber::class,
-            EvilNode::wrap(new Node\Scalar\LNumber(1)),
+            MutatedNode::wrap(new Node\Scalar\LNumber(1)),
             -1,
             [],
             $nominalAttributes,
@@ -126,7 +126,7 @@ final class MutationTest extends TestCase
             Plus::getName(),
             $nominalAttributes,
             Node\Scalar\LNumber::class,
-            EvilNode::wrap(new Node\Scalar\LNumber(1)),
+            MutatedNode::wrap(new Node\Scalar\LNumber(1)),
             0,
             [
                 CoverageLineData::with(
@@ -149,7 +149,7 @@ final class MutationTest extends TestCase
             Plus::getName(),
             $nominalAttributes,
             Node\Scalar\LNumber::class,
-            EvilNode::wrap(new Node\Scalar\LNumber(1)),
+            MutatedNode::wrap(new Node\Scalar\LNumber(1)),
             99,
             [
                 CoverageLineData::with(
@@ -172,7 +172,7 @@ final class MutationTest extends TestCase
             Plus::getName(),
             array_merge($nominalAttributes, ['foo' => 100, 'bar' => 1000]),
             Node\Scalar\LNumber::class,
-            EvilNode::wrap(new Node\Scalar\LNumber(1)),
+            MutatedNode::wrap(new Node\Scalar\LNumber(1)),
             0,
             [
                 CoverageLineData::with(
@@ -195,7 +195,7 @@ final class MutationTest extends TestCase
             Plus::getName(),
             $nominalAttributes,
             Node\Scalar\LNumber::class,
-            EvilNode::wrap(new Node\Scalar\LNumber(1)),
+            MutatedNode::wrap(new Node\Scalar\LNumber(1)),
             0,
             [],
             $nominalAttributes,
@@ -212,7 +212,7 @@ final class MutationTest extends TestCase
             Plus::getName(),
             $nominalAttributes,
             Node\Scalar\LNumber::class,
-            EvilNode::wrap([
+            MutatedNode::wrap([
                 new Node\Scalar\LNumber(1),
                 new Node\Scalar\LNumber(-1),
             ]),
