+++ tests/phpunit/Visitor/MutatorVisitorTest.php
@@ -36,7 +36,7 @@ declare(strict_types=1);
 namespace Infection\Tests\Visitor;
 
 use Generator;
-use Infection\EvilNode;
+use Infection\MutatedNode;
 use Infection\Mutation;
 use Infection\Mutator\FunctionSignature\PublicVisibility;
 use Infection\Tests\StringNormalizer;
@@ -118,7 +118,7 @@ PHP
                         'endFilePos' => -1,
                     ],
                     ClassMethod::class,
-                    EvilNode::wrap(new Nop()),
+                    MutatedNode::wrap(new Nop()),
                     0,
                     []
                 ),
@@ -170,7 +170,7 @@ PHP
                         'endFilePos' => -1,
                     ],
                     ClassMethod::class,
-                    EvilNode::wrap([new Nop(), new Nop()]),
+                    MutatedNode::wrap([new Nop(), new Nop()]),
                     0,
                     []
                 ),
@@ -224,7 +224,7 @@ PHP
                         'endFilePos' => -1,
                     ],
                     ClassMethod::class,
-                    EvilNode::wrap(new Nop()),
+                    MutatedNode::wrap(new Nop()),
                     0,
                     []
                 ),
@@ -292,7 +292,7 @@ PHP
                         'endFilePos' => -1,
                     ],
                     PublicVisibility::getName(),
-                    EvilNode::wrap(new Nop()),
+                    MutatedNode::wrap(new Nop()),
                     0,
                     []
                 ),
